Add kweb-benchmarks script
This is a python script that can be used to run various tasks related to
benchmarking kweb.  Currently the only task it knows how to execute is to
generate some useful graphs from the data generated by a modified httperf with
an additional --dump-events flag. We can extend this with more useful tasks as
we come up with them.

The graphs are generated via matplotlib and require numpy, wo you will have to
install these dependencies if you want to use this functionality.