Dirty hack options, mimicing kweb-pth
The options are #defines in os.h.  Toggle as you see fit.  It looks like
the webserver does worse when it has more I/O concurrency, probably due
to kernel issues.

The global alarms are a pain.  Should poll on those with separate
tchains, similar to how the kernel's alarm code works.